+++ tests/solver/TSTsstack_t.cc
@@ -44,8 +44,8 @@ TEST_F (SstackFixture, PushSstack) {
 
                 // all actions have the same lambda function which actually does
                 // nothing and receives a null index and two random numbers
-                frame.push (action_t {[] (size_t index, int val1, int val2) {},
-                    0, rand (), rand ()});
+                frame.push (action_t {[] (size_t index, size_t val1, size_t val2) {},
+                    0, size_t(rand ()), size_t(rand ())});
             }
 
             // in passing, why not, verify that the size of the frame is the
@@ -98,9 +98,9 @@ TEST_F (SstackFixture, UnwindSstack) {
             // all actions have the same lambda function which adds two numbers
             // and receives the arguments to add ---note the index is uselss in
             // this tests
-            frame0.push (action_t {[] (size_t index, int val1, int val2) {
+            frame0.push (action_t {[] (size_t index, size_t val1, size_t val2) {
                 sum += (val1 + val2);
-            }, 0, ints[j*2], ints[j*2+1]});
+            }, 0, size_t(ints[j*2]), size_t(ints[j*2+1])});
         }
 
         // push this frame onto the stack
@@ -117,10 +117,10 @@ TEST_F (SstackFixture, UnwindSstack) {
             // all actions have the same lambda function which inserts two
             // numbers (in reversed order) in a vector and receives the
             // arguments to insert ---note the index is uselss in this tests
-            frame1.push (action_t {[] (size_t index, int val1, int val2) {
+            frame1.push (action_t {[] (size_t index, size_t val1, size_t val2) {
                 stackContainer.push_back (val2);
                 stackContainer.push_back (val1);
-            }, 0, ints[j*2], ints[j*2+1]});
+            }, 0, size_t(ints[j*2]), size_t(ints[j*2+1])});
         }
 
         // push this frame onto the stack
